ESN Layers

Input Layers

ReservoirComputing.WeightedLayer — Type

WeightedInput(scaling)
WeightedInput(;scaling=0.1)

Creates a WeightedInput layer initializer for Echo State Networks. This initializer generates a weighted input matrix with random non-zero elements distributed uniformly within the range [-scaling, scaling], following the approach in [Lu].

Parameters

scaling: The scaling factor for the weight distribution (default: 0.1).

Returns

A WeightedInput instance to be used for initializing the input layer of an ESN.

Reference: [Lu]: Lu, Zhixin, et al. "Reservoir observers: Model-free inference of unmeasured variables in chaotic systems." Chaos: An Interdisciplinary Journal of Nonlinear Science 27.4 (2017): 041102.

ReservoirComputing.DenseLayer — Type

DenseLayer(scaling)
DenseLayer(;scaling=0.1)

Creates a DenseLayer initializer for Echo State Networks, generating a fully connected input layer. The layer is initialized with random weights uniformly distributed within [-scaling, scaling]. This scaling factor can be provided either as an argument or a keyword argument. The DenseLayer is the default input layer in ESN construction.

Parameters

scaling: The scaling factor for weight distribution (default: 0.1).

Returns

A DenseLayer instance for initializing the ESN's input layer.

ReservoirComputing.SparseLayer — Type

SparseLayer(scaling, sparsity)
SparseLayer(scaling; sparsity=0.1)
SparseLayer(;scaling=0.1, sparsity=0.1)

Creates a SparseLayer initializer for Echo State Networks, generating a sparse input layer. The layer is initialized with weights distributed within [-scaling, scaling] and a specified sparsity level. Both scaling and sparsity can be set as arguments or keyword arguments.

Parameters

scaling: Scaling factor for weight distribution (default: 0.1).

sparsity: Sparsity level of the layer (default: 0.1).

Returns

A SparseLayer instance for initializing ESN's input layer with sparse connections.

ReservoirComputing.InformedLayer — Type

InformedLayer(model_in_size; scaling=0.1, gamma=0.5)

Creates an InformedLayer initializer for Echo State Networks (ESNs) that generates a weighted input layer matrix. The matrix contains random non-zero elements drawn from the range [-scaling, scaling]. This initializer ensures that a fraction (gamma) of reservoir nodes are exclusively connected to the raw inputs, while the rest are connected to the outputs of a prior knowledge model, as described in [Pathak].

Arguments

model_in_size: The size of the prior knowledge model's output, which determines the number of columns in the input layer matrix.

Keyword Arguments

scaling: The absolute value of the weights (default: 0.1).

gamma: The fraction of reservoir nodes connected exclusively to raw inputs (default: 0.5).

Returns

An InformedLayer instance for initializing the ESN's input layer matrix.

Reference: [Pathak]: Jaideep Pathak et al. "Hybrid Forecasting of Chaotic Processes: Using Machine Learning in Conjunction with a Knowledge-Based Model" (2018).

ReservoirComputing.MinimumLayer — Type

MinimumLayer(weight, sampling)
MinimumLayer(weight; sampling=BernoulliSample(0.5))
MinimumLayer(;weight=0.1, sampling=BernoulliSample(0.5))

Creates a MinimumLayer initializer for Echo State Networks, generating a fully connected input layer. This layer has a uniform absolute weight value (weight) with the sign of each weight determined by the sampling method. This approach, as detailed in [Rodan1] and [Rodan2], allows for controlled weight distribution in the layer.

Parameters

weight: Absolute value of weights in the layer.

sampling: Method for determining the sign of weights (default: BernoulliSample(0.5)).

Returns

A MinimumLayer instance for initializing the ESN's input layer.

References: [Rodan1]: Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE Transactions on Neural Networks 22.1 (2010): 131-144. [Rodan2]: Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs with regular jumps." Neural Computation 24.7 (2012): 1822-1852.

ReservoirComputing.NullLayer — Type

NullLayer()

Creates a NullLayer initializer for Echo State Networks (ESNs) that generates a vector of zeros.

Returns

A NullLayer instance for initializing the ESN's input layer matrix.

The signs in the MinimumLayer are chosen based on the following methods:

ReservoirComputing.BernoulliSample — Type

BernoulliSample(p)
BernoulliSample(;p=0.5)

Creates a BernoulliSample constructor for the MinimumLayer. It uses a Bernoulli distribution to determine the sign of weights in the input layer. The parameter p sets the probability of a weight being positive, as per the Distributions package. This method of sign weight determination for input layers is based on the approach in [Rodan].

Parameters

p: Probability of a positive weight (default: 0.5).

Returns

A BernoulliSample instance for generating sign weights in MinimumLayer.

Reference: [Rodan]: Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE Transactions on Neural Networks 22.1 (2010): 131-144.

ReservoirComputing.IrrationalSample — Type

IrrationalSample(irrational, start)
IrrationalSample(;irrational=pi, start=1)

Creates an IrrationalSample constructor for the MinimumLayer. It determines the sign of weights in the input layer based on the decimal expansion of an irrational number. The start parameter sets the starting point in the decimal sequence. The signs are assigned based on the thresholding of each decimal digit against 4.5, as described in [Rodan].

Parameters

irrational: An irrational number for weight sign determination (default: π).

start: Starting index in the decimal expansion (default: 1).

Returns

An IrrationalSample instance for generating sign weights in MinimumLayer.

Reference: [Rodan]: Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs with regular jumps." Neural Computation 24.7 (2012): 1822-1852.

To derive the matrix one can call the following function:

ReservoirComputing.create_layer — Function

create_layer(input_layer::AbstractLayer, res_size, in_size)

Generates a matrix layer of size res_size x in_size, constructed according to the specifications of the input_layer.

Parameters

input_layer: An instance of AbstractLayer determining the layer construction.

res_size: The number of rows (reservoir size) for the layer.

in_size: The number of columns (input size) for the layer.

Returns

A matrix representing the constructed layer.

To create new input layers, it suffices to define a new struct containing the needed parameters of the new input layer. This struct will need to be an AbstractLayer, so the create_layer function can be dispatched over it. The workflow should follow this snippet:

#creation of the new struct for the layer
struct MyNewLayer <: AbstractLayer
    #the layer params go here
end

#dispatch over the function to actually build the layer matrix
function create_layer(input_layer::MyNewLayer, res_size, in_size)
    #the new algorithm to build the input layer goes here
end

Reservoirs

ReservoirComputing.RandSparseReservoir — Type

RandSparseReservoir(res_size, radius, sparsity)
RandSparseReservoir(res_size; radius=1.0, sparsity=0.1)

Returns a random sparse reservoir initializer, which generates a matrix of size res_size x res_size with the specified sparsity and scaled spectral radius according to radius. This type of reservoir initializer is commonly used in Echo State Networks (ESNs) for capturing complex temporal dependencies.

Arguments

res_size: The size of the reservoir matrix.

radius: The desired spectral radius of the reservoir. By default, it is set to 1.0.

sparsity: The sparsity level of the reservoir matrix, controlling the fraction of zero elements. By default, it is set to 0.1.

Returns

A RandSparseReservoir object that can be used as a reservoir initializer in ESN construction.

References

This type of reservoir initialization is a common choice in ESN construction for its ability to capture temporal dependencies in data. However, there is no specific reference associated with this function.

ReservoirComputing.PseudoSVDReservoir — Type

PseudoSVDReservoir(max_value, sparsity, sorted, reverse_sort)
PseudoSVDReservoir(max_value, sparsity; sorted=true, reverse_sort=false)

Returns an initializer to build a sparse reservoir matrix with the given sparsity by using a pseudo-SVD approach as described in [yang].

Arguments

res_size: The size of the reservoir matrix.

max_value: The maximum absolute value of elements in the matrix.

sparsity: The desired sparsity level of the reservoir matrix.

sorted: A boolean indicating whether to sort the singular values before creating the diagonal matrix. By default, it is set to true.

reverse_sort: A boolean indicating whether to reverse the sorted singular values. By default, it is set to false.

Returns

A PseudoSVDReservoir object that can be used as a reservoir initializer in ESN construction.

References

This reservoir initialization method, based on a pseudo-SVD approach, is inspired by the work in [yang], which focuses on designing polynomial echo state networks for time series prediction.

ReservoirComputing.DelayLineReservoir — Type

DelayLineReservoir(res_size, weight)
DelayLineReservoir(res_size; weight=0.1)

Returns a Delay Line Reservoir matrix constructor to obtain a deterministic reservoir as described in [Rodan2010].

Arguments

res_size::Int: The size of the reservoir.

weight::T: The weight determines the absolute value of all the connections in the reservoir.

Returns

A DelayLineReservoir object.

References

IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.DelayLineBackwardReservoir — Type

DelayLineBackwardReservoir(res_size, weight, fb_weight)
DelayLineBackwardReservoir(res_size; weight=0.1, fb_weight=0.2)

Returns a Delay Line Reservoir constructor to create a matrix with backward connections as described in [Rodan2010]. The weight and fb_weight can be passed as either arguments or keyword arguments, and they determine the absolute values of the connections in the reservoir.

Arguments

res_size::Int: The size of the reservoir.

weight::T: The weight determines the absolute value of forward connections in the reservoir.

fb_weight::T: The fb_weight determines the absolute value of backward connections in the reservoir.

Returns

A DelayLineBackwardReservoir object.

References

IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.SimpleCycleReservoir — Type

SimpleCycleReservoir(res_size, weight)
SimpleCycleReservoir(res_size; weight=0.1)

Returns a Simple Cycle Reservoir constructor to build a reservoir matrix as described in [Rodan2010]. The weight can be passed as an argument or a keyword argument, and it determines the absolute value of all the connections in the reservoir.

Arguments

res_size::Int: The size of the reservoir.

weight::T: The weight determines the absolute value of connections in the reservoir.

Returns

A SimpleCycleReservoir object.

References

IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.CycleJumpsReservoir — Type

CycleJumpsReservoir(res_size; cycle_weight=0.1, jump_weight=0.1, jump_size=3)
CycleJumpsReservoir(res_size, cycle_weight, jump_weight, jump_size)

Return a Cycle Reservoir with Jumps constructor to create a reservoir matrix as described in [Rodan2012]. The cycle_weight, jump_weight, and jump_size can be passed as arguments or keyword arguments, and they determine the absolute values of connections in the reservoir. The jump_size determines the jumps between jump_weights.

Arguments

res_size::Int: The size of the reservoir.

cycle_weight::T: The weight of cycle connections.

jump_weight::T: The weight of jump connections.

jump_size::Int: The number of steps between jump connections.

Returns

A CycleJumpsReservoir object.

References

with regular jumps." Neural computation 24.7 (2012): 1822-1852.

ReservoirComputing.NullReservoir — Type

NullReservoir()

Return a constructor for a matrix of zeros with dimensions res_size x res_size.

Arguments

None

Returns

A NullReservoir object.

References

None

Like for the input layers, to actually build the matrix of the reservoir, one can call the following function:

ReservoirComputing.create_reservoir — Function

create_reservoir(reservoir::AbstractReservoir, res_size)
create_reservoir(reservoir, args...)

Given an AbstractReservoir constructor and the size of the reservoir (res_size), this function returns the corresponding reservoir matrix. Alternatively, it accepts a pre-generated matrix.

Arguments

reservoir: An AbstractReservoir object or constructor.

res_size: The size of the reservoir matrix.

matrix_type: The type of the resulting matrix. By default, it is set to Matrix{Float64}.

Returns

A matrix representing the reservoir, generated based on the properties of the specified reservoir object or constructor.

References

The choice of reservoir initialization is crucial in Echo State Networks (ESNs) for achieving effective temporal modeling. Specific references for reservoir initialization methods may vary based on the type of reservoir used, but the practice of initializing reservoirs for ESNs is widely documented in the ESN literature.

To create a new reservoir, the procedure is similar to the one for the input layers. First, the definition of the new struct of type AbstractReservoir with the reservoir parameters is needed. Then the dispatch over the create_reservoir function makes the model actually build the reservoir matrix. An example of the workflow is given in the following snippet:

#creation of the new struct for the reservoir
struct MyNewReservoir <: AbstractReservoir
    #the reservoir params go here
end

#dispatch over the function to build the reservoir matrix
function create_reservoir(reservoir::AbstractReservoir, res_size)
    #the new algorithm to build the reservoir matrix goes here
end

yangYang, Cuili, et al. "Design of polynomial echo state networks for time series prediction." Neurocomputing 290 (2018): 148-160.

Rodan2010Rodan, Ali, and Peter Tino. "Minimum complexity echo state network."

Rodan2010Rodan, Ali, and Peter Tino. "Minimum complexity echo state network."

Rodan2010Rodan, Ali, and Peter Tino. "Minimum complexity echo state network."

Rodan2012Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs