Echo State Networks Initializers

Input layers

ReservoirComputing.scaled_rand — Function

scaled_rand(rng::AbstractRNG, ::Type{T}, dims::Integer...; scaling=T(0.1)) where {T <: Number}

Create and return a matrix with random values, uniformly distributed within a range defined by scaling. This function is useful for initializing matrices, such as the layers of a neural network, with scaled random values.

Arguments

rng: An instance of AbstractRNG for random number generation.

T: The data type for the elements of the matrix.

dims: Dimensions of the matrix. It must be a 2-element tuple specifying the number of rows and columns (e.g., (res_size, in_size)).

scaling: A scaling factor to define the range of the uniform distribution. The matrix elements will be randomly chosen from the range [-scaling, scaling]. Defaults to T(0.1).

Returns

A matrix of type with dimensions specified by dims. Each element of the matrix is a random number uniformly distributed between -scaling and scaling.

Example

rng = Random.default_rng()
matrix = scaled_rand(rng, Float64, (100, 50); scaling=0.2)

ReservoirComputing.weighted_init — Function

weighted_init(rng::AbstractRNG, ::Type{T}, dims::Integer...; scaling=T(0.1)) where {T <: Number}

Create and return a matrix representing a weighted input layer for Echo State Networks (ESNs). This initializer generates a weighted input matrix with random non-zero elements distributed uniformly within the range [-scaling, scaling], inspired by the approach in [Lu].

Arguments

rng: An instance of AbstractRNG for random number generation.

T: The data type for the elements of the matrix.

dims: A 2-element tuple specifying the approximate reservoir size and input size (e.g., (approx_res_size, in_size)).

scaling: The scaling factor for the weight distribution. Defaults to T(0.1).

Returns

A matrix representing the weighted input layer as defined in [Lu2017]. The matrix dimensions will be adjusted to ensure each input unit connects to an equal number of reservoir units.

Example

rng = Random.default_rng()
input_layer = weighted_init(rng, Float64, (3, 300); scaling=0.2)

References

ReservoirComputing.informed_init — Function

informed_init(rng::AbstractRNG, ::Type{T}, dims::Integer...; scaling=T(0.1), model_in_size, gamma=T(0.5)) where {T <: Number}

Create a layer of a neural network.

Arguments

rng::AbstractRNG: The random number generator.

T::Type: The data type.

dims::Integer...: The dimensions of the layer.

scaling::T = T(0.1): The scaling factor for the input matrix.

model_in_size: The size of the input model.

gamma::T = T(0.5): The gamma value.

Returns

input_matrix: The created input matrix for the layer.

Example

rng = Random.default_rng()
dims = (100, 200)
model_in_size = 50
input_matrix = informed_init(rng, Float64, dims; model_in_size=model_in_size)

ReservoirComputing.minimal_init — Function

irrational_sample_init(rng::AbstractRNG, ::Type{T}, dims::Integer...;
weight = 0.1,
sampling = IrrationalSample(; irrational = pi, start = 1)
) where {T <: Number}

Create a layer matrix using the provided random number generator and sampling parameters.

Arguments

rng::AbstractRNG: The random number generator used to generate random numbers.

dims::Integer...: The dimensions of the layer matrix.

weight: The weight used to fill the layer matrix. Default is 0.1.

sampling: The sampling parameters used to generate the input matrix. Default is IrrationalSample(irrational = pi, start = 1).

Returns

The layer matrix generated using the provided random number generator and sampling parameters.

Example

using Random
rng = Random.default_rng()
dims = (3, 2)
weight = 0.5
layer_matrix = irrational_sample_init(rng, Float64, dims; weight=weight,
    sampling=IrrationalSample(; irrational=sqrt(2), start=1))

Reservoirs

ReservoirComputing.rand_sparse — Function

rand_sparse(rng::AbstractRNG, ::Type{T}, dims::Integer...; radius=1.0, sparsity=0.1)

Create and return a random sparse reservoir matrix for use in Echo State Networks (ESNs). The matrix will be of size specified by dims, with specified sparsity and scaled spectral radius according to radius.

Arguments

rng: An instance of AbstractRNG for random number generation.

T: The data type for the elements of the matrix.

dims: Dimensions of the reservoir matrix.

radius: The desired spectral radius of the reservoir. Defaults to 1.0.

sparsity: The sparsity level of the reservoir matrix, controlling the fraction of zero elements. Defaults to 0.1.

Returns

A matrix representing the random sparse reservoir.

References

This type of reservoir initialization is commonly used in ESNs for capturing temporal dependencies in data.

ReservoirComputing.delay_line — Function

delay_line(rng::AbstractRNG, ::Type{T}, dims::Integer...; weight=0.1) where {T <: Number}

Create and return a delay line reservoir matrix for use in Echo State Networks (ESNs). A delay line reservoir is a deterministic structure where each unit is connected only to its immediate predecessor with a specified weight. This method is particularly useful for tasks that require specific temporal processing.

Arguments

rng: An instance of AbstractRNG for random number generation. This argument is not used in the current implementation but is included for consistency with other initialization functions.

T: The data type for the elements of the matrix.

dims: Dimensions of the reservoir matrix. Typically, this should be a tuple of two equal integers representing a square matrix.

weight: The weight determines the absolute value of all connections in the reservoir. Defaults to 0.1.

Returns

A delay line reservoir matrix with dimensions specified by dims. The matrix is initialized such that each element in the i+1th row and ith column is set to weight, and all other elements are zeros.

Example

reservoir = delay_line(Float64, 100, 100; weight=0.2)

References

This type of reservoir initialization is described in: Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE Transactions on Neural Networks 22.1 (2010): 131-144.

ReservoirComputing.delay_line_backward — Function

delay_line_backward(rng::AbstractRNG, ::Type{T}, dims::Integer...;
    weight = T(0.1), fb_weight = T(0.2)) where {T <: Number}

Create a delay line backward reservoir with the specified by dims and weights. Creates a matrix with backward connections as described in [Rodan2010]. The weight and fb_weight can be passed as either arguments or keyword arguments, and they determine the absolute values of the connections in the reservoir.

Arguments

rng::AbstractRNG: Random number generator.

T::Type: Type of the elements in the reservoir matrix.

dims::Integer...: Dimensions of the reservoir matrix.

weight::T: The weight determines the absolute value of forward connections in the reservoir, and is set to 0.1 by default.

fb_weight::T: The fb_weight determines the absolute value of backward connections in the reservoir, and is set to 0.2 by default.

Returns

Reservoir matrix with the dimensions specified by dims and weights.

References

ReservoirComputing.cycle_jumps — Function

cycle_jumps(rng::AbstractRNG, ::Type{T}, dims::Integer...;
    cycle_weight = T(0.1), jump_weight = T(0.1), jump_size = 3) where {T <: Number}

Create a cycle jumps reservoir with the specified dimensions, cycle weight, jump weight, and jump size.

Arguments

rng::AbstractRNG: Random number generator.

T::Type: Type of the elements in the reservoir matrix.

dims::Integer...: Dimensions of the reservoir matrix.

cycle_weight::T = T(0.1): The weight of cycle connections.

jump_weight::T = T(0.1): The weight of jump connections.

jump_size::Int = 3: The number of steps between jump connections.

Returns

Reservoir matrix with the specified dimensions, cycle weight, jump weight, and jump size.

References

ReservoirComputing.simple_cycle — Function

simple_cycle(rng::AbstractRNG, ::Type{T}, dims::Integer...;
    weight = T(0.1)) where {T <: Number}

Create a simple cycle reservoir with the specified dimensions and weight.

Arguments

rng::AbstractRNG: Random number generator.

T::Type: Type of the elements in the reservoir matrix.

dims::Integer...: Dimensions of the reservoir matrix.

weight::T = T(0.1): Weight of the connections in the reservoir matrix.

Returns

Reservoir matrix with the dimensions specified by dims and weights.

References

ReservoirComputing.pseudo_svd — Function

pseudo_svd(rng::AbstractRNG, ::Type{T}, dims::Integer...; 
    max_value, sparsity, sorted = true, reverse_sort = false) where {T <: Number}

Returns an initializer to build a sparse reservoir matrix with the given `sparsity` by using a pseudo-SVD approach as described in [^yang].

Arguments

rng::AbstractRNG: Random number generator.

T::Type: Type of the elements in the reservoir matrix.

dims::Integer...: Dimensions of the reservoir matrix.

max_value: The maximum absolute value of elements in the matrix.

sparsity: The desired sparsity level of the reservoir matrix.

sorted: A boolean indicating whether to sort the singular values before creating the diagonal matrix. By default, it is set to true.

reverse_sort: A boolean indicating whether to reverse the sorted singular values. By default, it is set to false.

Returns

Reservoir matrix with the specified dimensions, max value, and sparsity.

References

This reservoir initialization method, based on a pseudo-SVD approach, is inspired by the work in [yang], which focuses on designing polynomial echo state networks for time series prediction.

Lu2017Lu, Zhixin, et al. "Reservoir observers: Model-free inference of unmeasured variables in chaotic systems." Chaos: An Interdisciplinary Journal of Nonlinear Science 27.4 (2017): 041102.

Rodan2010Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

Rodan2012Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs with regular jumps." Neural computation 24.7 (2012): 1822-1852.

Rodan2010Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

yangYang, Cuili, et al. "Design of polynomial echo state networks for time series prediction." Neurocomputing 290 (2018): 148-160.