delay_line

ReservoirComputing.delay_line — Function

delay_line([rng], [T], dims...;
    delay_weight=0.1, delay_shift=1,
    return_sparse=false, radius=nothing, kwargs...)

Create and return a delay line reservoir matrix (Rodan and Tino, 2011).

\[W_{i,j} = \begin{cases} r, & \text{if } i = j + 1, j \in [1, D_{\mathrm{res}} - 1], \\[6pt] 0, & \text{otherwise.} \end{cases}\]

Arguments

rng: Random number generator. Default is Utils.default_rng()from WeightInitializers.

T: Type of the elements in the reservoir matrix. Default is Float32.

dims: Dimensions of the reservoir matrix.

Keyword arguments

delay_weight: Determines the value of all connections in the reservoir. This can be provided as a single value or an array. In case it is provided as an array please make sure that the length of the array matches the length of the sub-diagonal you want to populate. Default is 0.1.

delay_shift: delay line shift. Default is 1.

radius: The desired spectral radius of the reservoir. If nothing is passed, no scaling takes place. Defaults to nothing.

return_sparse: flag for returning a sparse matrix. true requires SparseArrays to be loaded. Default is false.

sampling_type: Sampling that decides the distribution of weight negative numbers. If set to :no_sample the sign is unchanged. If set to :bernoulli_sample! then each weight can be positive with a probability set by positive_prob. If set to :irrational_sample! the weight is negative if the decimal number of the irrational number chosen is odd. If set to :regular_sample!, each weight will be assigned a negative sign after the chosen strides. strides can be a single number or an array. Default is :no_sample.

positive_prob: probability of the weight being positive when sampling_type is set to :bernoulli_sample!. Default is 0.5.

irrational: Irrational number whose decimals decide the sign of weight. Default is pi.

start: Which place after the decimal point the counting starts for the irrational sign counting. Default is 1.

strides: number of strides for assigning negative value to a weight. It can be an integer or an array. Default is 2.

Examples

Default call:

julia> res_matrix = delay_line(5, 5)
5×5 Matrix{Float32}:
 0.0  0.0  0.0  0.0  0.0
 0.1  0.0  0.0  0.0  0.0
 0.0  0.1  0.0  0.0  0.0
 0.0  0.0  0.1  0.0  0.0
 0.0  0.0  0.0  0.1  0.0

Changing weights:

julia> res_matrix = delay_line(5, 5; delay_weight = 1)
5×5 Matrix{Float32}:
 0.0  0.0  0.0  0.0  0.0
 1.0  0.0  0.0  0.0  0.0
 0.0  1.0  0.0  0.0  0.0
 0.0  0.0  1.0  0.0  0.0
 0.0  0.0  0.0  1.0  0.0

Changing weights to a custom array:

julia> res_matrix = delay_line(5, 5; delay_weight = rand(Float32, 4))
5×5 Matrix{Float32}:
 0.0       0.0       0.0      0.0        0.0
 0.398408  0.0       0.0      0.0        0.0
 0.0       0.624473  0.0      0.0        0.0
 0.0       0.0       0.66302  0.0        0.0
 0.0       0.0       0.0      0.0780818  0.0

Changing sign of the weights with different samplings:

julia> res_matrix = delay_line(5, 5; sampling_type=:irrational_sample!)
5×5 Matrix{Float32}:
 0.0  0.0   0.0   0.0  0.0
 -0.1  0.0   0.0   0.0  0.0
 0.0  0.1   0.0   0.0  0.0
 0.0  0.0  -0.1   0.0  0.0
 0.0  0.0   0.0  -0.1  0.0

julia> res_matrix = delay_line(5, 5; sampling_type=:bernoulli_sample!)
5×5 Matrix{Float32}:
 0.0   0.0  0.0   0.0  0.0
 0.1   0.0  0.0   0.0  0.0
 0.0  -0.1  0.0   0.0  0.0
 0.0   0.0  0.1   0.0  0.0
 0.0   0.0  0.0  -0.1  0.0

Shifting the delay line:

julia> res_matrix = delay_line(5, 5; delay_shift = 3)
5×5 Matrix{Float32}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.1  0.0  0.0  0.0  0.0
 0.0  0.1  0.0  0.0  0.0

Returning as sparse:

julia> using SparseArrays

julia> res_matrix = delay_line(5, 5; return_sparse=true)
5×5 SparseMatrixCSC{Float32, Int64} with 4 stored entries:
  ⋅    ⋅    ⋅    ⋅    ⋅
 0.1   ⋅    ⋅    ⋅    ⋅
  ⋅   0.1   ⋅    ⋅    ⋅
  ⋅    ⋅   0.1   ⋅    ⋅
  ⋅    ⋅    ⋅   0.1   ⋅

References

Rodan, A. and Tino, P. (2011). Minimum Complexity Echo State Network. IEEE Transactions on Neural Networks 22, 131–144.