forward_connection

ReservoirComputing.forward_connection — Function

forward_connection([rng], [T], dims...;
    forward_weight=0.1, radius=nothing, return_sparse=false,
    kwargs...)

Creates a reservoir based on a forward connection of weights (Elsarraj et al., 2019).

This architecture is referred to as TP5 in the original paper.

\[W_{i,j} = \begin{cases} r, & \text{if } j = i - 2 \text{ for } i = 3 \dots N \\ 0, & \text{otherwise} \end{cases}\]

Arguments

rng: Random number generator. Default is Utils.default_rng()from WeightInitializers.

T: Type of the elements in the reservoir matrix. Default is Float32.

dims: Dimensions of the reservoir matrix.

Keyword arguments

forward_weight: Weight of the cycle connections in the reservoir matrix. This can be provided as a single value or an array. In case it is provided as an array please make sure that the length of the array matches the length of the sub-diagonal you want to populate. Default is 0.1.

radius: The desired spectral radius of the reservoir. If nothing is passed, no scaling takes place. Defaults to nothing.

return_sparse: flag for returning a sparse matrix. true requires SparseArrays to be loaded. Default is false.

sampling_type: Sampling that decides the distribution of forward_weight negative numbers. If set to :no_sample the sign is unchanged. If set to :bernoulli_sample! then each forward_weight can be positive with a probability set by positive_prob. If set to :irrational_sample! the forward_weight is negative if the decimal number of the irrational number chosen is odd. If set to :regular_sample!, each weight will be assigned a negative sign after the chosen strides. strides can be a single number or an array. Default is :no_sample.

positive_prob: probability of the forward_weight being positive when sampling_type is set to :bernoulli_sample!. Default is 0.5.

irrational: Irrational number whose decimals decide the sign of forward_weight. Default is pi.

start: Which place after the decimal point the counting starts for the irrational sign counting. Default is 1.

strides: number of strides for assigning negative value to a weight. It can be an integer or an array. Default is 2.

Examples

Default kwargs:

julia> reservoir_matrix = forward_connection(5, 5)
5×5 Matrix{Float32}:
 0.0  0.0  0.0  0.0  0.0
 0.0  0.0  0.0  0.0  0.0
 0.1  0.0  0.0  0.0  0.0
 0.0  0.1  0.0  0.0  0.0
 0.0  0.0  0.1  0.0  0.0

Changing the weights magnitudes to a different unique value:

julia> forward_connection(5, 5; forward_weight=0.99)
5×5 Matrix{Float32}:
 0.0   0.0   0.0   0.0  0.0
 0.0   0.0   0.0   0.0  0.0
 0.99  0.0   0.0   0.0  0.0
 0.0   0.99  0.0   0.0  0.0
 0.0   0.0   0.99  0.0  0.0

Changing the weights signs with different sampling techniques:

julia> forward_connection(5, 5; sampling_type=:irrational_sample!)
5×5 Matrix{Float32}:
  0.0  0.0   0.0  0.0  0.0
  0.0  0.0   0.0  0.0  0.0
 -0.1  0.0   0.0  0.0  0.0
  0.0  0.1   0.0  0.0  0.0
  0.0  0.0  -0.1  0.0  0.0

julia> forward_connection(5, 5; sampling_type=:irrational_sample!)
5×5 Matrix{Float32}:
  0.0  0.0   0.0  0.0  0.0
  0.0  0.0   0.0  0.0  0.0
  -0.1  0.0   0.0  0.0  0.0
  0.0  0.1   0.0  0.0  0.0
  0.0  0.0  -0.1  0.0  0.0

Changing the weights to random numbers. Note that the length of the given array must be at least as long as the subdiagonal one wants to fill:

julia> reservoir_matrix = forward_connection(5, 5; forward_weight=rand(Float32, 3))
5×5 Matrix{Float32}:
 0.0       0.0       0.0       0.0  0.0
 0.0       0.0       0.0       0.0  0.0
 0.274221  0.0       0.0       0.0  0.0
 0.0       0.111511  0.0       0.0  0.0
 0.0       0.0       0.618345  0.0  0.0

Returning a sparse matrix:

julia> reservoir_matrix = forward_connection(10, 10; return_sparse=true)
10×10 SparseMatrixCSC{Float32, Int64} with 8 stored entries:
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
 0.1   ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅   0.1   ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅   0.1   ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅   0.1   ⋅    ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅   0.1   ⋅    ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅   0.1   ⋅    ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅   0.1   ⋅    ⋅    ⋅
  ⋅    ⋅    ⋅    ⋅    ⋅    ⋅    ⋅   0.1   ⋅    ⋅

References

Elsarraj, D.; Qisi, M. A.; Rodan, A.; Obeid, N.; Sharieh, A. and Faris, H. (2019). Demystifying echo state network with deterministic simple topologies. International Journal of Computational Science and Engineering 19, 407–417.