selfloopdelaylinebackward

ReservoirComputing.selfloop_delayline_backward — Function

selfloop_delayline_backward([rng], [T], dims...;
    delay_weight=0.1, selfloop_weight=0.1, fb_weight=0.1,
    fb_shift=2, delya_shift=1, radius=nothing, return_sparse=false,
    fb_kwargs=(), selfloop_kwargs=(), delay_kwargs=())

Creates a reservoir based on a delay line with the addition of self loops and backward connections shifted by one (Elsarraj et al., 2019).

This architecture is referred to as TP3 in the original paper.

\[W_{i,j} = \begin{cases} ll, & \text{if } i = j \text{ for } i = 1 \dots N \\ r, & \text{if } j = i - 1 \text{ for } i = 2 \dots N \\ r, & \text{if } j = i - 2 \text{ for } i = 3 \dots N \\ 0, & \text{otherwise} \end{cases}\]

Arguments

rng: Random number generator. Default is Utils.default_rng()from WeightInitializers.

T: Type of the elements in the reservoir matrix. Default is Float32.

dims: Dimensions of the reservoir matrix.

Keyword arguments

delay_weight: Weight of the delay line connections in the reservoir matrix. This can be provided as a single value or an array. In case it is provided as an array please make sure that the length of the array matches the length of the cycle you want to populate. Default is 0.1.

selfloop_weight: Weight of the self loops in the reservoir matrix. This can be provided as a single value or an array. In case it is provided as an array please make sure that the length of the array matches the length of the diagonal you want to populate. Default is 0.1.

fb_weight: Weight of the feedback in the reservoir matrix. This can be provided as a single value or an array. In case it is provided as an array please make sure that the length of the array matches the length of the diagonal you want to populate. Default is 0.1.

fb_shift: How far the backward connection will be from the diagonal. Default is 1.

delay_shift: delay line shift relative to the diagonal. Default is 1.

radius: The desired spectral radius of the reservoir. If nothing is passed, no scaling takes place. Defaults to nothing.

return_sparse: flag for returning a sparse matrix. true requires SparseArrays to be loaded. Default is false.

delay_kwargs, selfloop_kwargs, and fb_kwargs: named tuples that control the kwargs for the weights generation. The kwargs are as follows:

sampling_type: Sampling that decides the distribution of weight negative numbers. If set to :no_sample the sign is unchanged. If set to :bernoulli_sample! then each weight can be positive with a probability set by positive_prob. If set to :irrational_sample! the weight is negative if the decimal number of the irrational number chosen is odd. If set to :regular_sample!, each weight will be assigned a negative sign after the chosen strides. strides can be a single number or an array. Default is :no_sample.

positive_prob: probability of the weight being positive when sampling_type is set to :bernoulli_sample!. Default is 0.5.

irrational: Irrational number whose decimals decide the sign of weight. Default is pi.

start: Which place after the decimal point the counting starts for the irrational sign counting. Default is 1.

strides: number of strides for assigning negative value to a weight. It can be an integer or an array. Default is 2.

Examples

Default call:

julia> res_matrix = selfloop_delayline_backward(5, 5)
5×5 Matrix{Float32}:
 0.1  0.0  0.1  0.0  0.0
 0.1  0.1  0.0  0.1  0.0
 0.0  0.1  0.1  0.0  0.1
 0.0  0.0  0.1  0.1  0.0
 0.0  0.0  0.0  0.1  0.1

Changing weights:

julia> res_matrix = selfloop_delayline_backward(5, 5; selfloop_weight=0.3, fb_weight=0.99, delay_weight=-0.5)
5×5 Matrix{Float32}:
  0.3   0.0   0.99   0.0   0.0
 -0.5   0.3   0.0    0.99  0.0
  0.0  -0.5   0.3    0.0   0.99
  0.0   0.0  -0.5    0.3   0.0
  0.0   0.0   0.0   -0.5   0.3

Changing weights to custom arrays:

julia> res_matrix = selfloop_delayline_backward(5, 5; selfloop_weight=randn(5), fb_weight=rand(5), delay_weight=-rand(5))
5×5 Matrix{Float32}:
 -1.22847    0.0       0.384073   0.0        0.0
 -0.699175   2.63937   0.0        0.345408   0.0
  0.0       -0.5171   -0.452312   0.0        0.0205082
  0.0        0.0      -0.193893   1.45921    0.0
  0.0        0.0       0.0       -0.453015  -1.43402

Changing sign of the weights with different samplings:

julia> res_matrix = selfloop_delayline_backward(5, 5; selfloop_kwargs=(;sampling_type=:irrational_sample!))
5×5 Matrix{Float32}:
 -0.1  0.0   0.1   0.0   0.0
  0.1  0.1   0.0   0.1   0.0
  0.0  0.1  -0.1   0.0   0.1
  0.0  0.0   0.1  -0.1   0.0
  0.0  0.0   0.0   0.1  -0.1

julia> res_matrix = selfloop_delayline_backward(5, 5; delay_kwargs=(;sampling_type=:bernoulli_sample!))
5×5 Matrix{Float32}:
 0.1   0.0  0.1   0.0  0.0
 0.1   0.1  0.0   0.1  0.0
 0.0  -0.1  0.1   0.0  0.1
 0.0   0.0  0.1   0.1  0.0
 0.0   0.0  0.0  -0.1  0.1

julia> res_matrix = selfloop_delayline_backward(5, 5; fb_kwargs=(;sampling_type=:regular_sample!))
5×5 Matrix{Float32}:
 0.1  0.0  0.1   0.0  0.0
 0.1  0.1  0.0  -0.1  0.0
 0.0  0.1  0.1   0.0  0.1
 0.0  0.0  0.1   0.1  0.0
 0.0  0.0  0.0   0.1  0.1

Shifting the delay and the backward line:

julia> res_matrix = selfloop_delayline_backward(5, 5; delay_shift=3, fb_shift=2)
5×5 Matrix{Float32}:
 0.1  0.0  0.1  0.0  0.0
 0.0  0.1  0.0  0.1  0.0
 0.0  0.0  0.1  0.0  0.1
 0.1  0.0  0.0  0.1  0.0
 0.0  0.1  0.0  0.0  0.1

Returning as sparse:

julia> using SparseArrays

julia> res_matrix = selfloop_delayline_backward(5, 5; return_sparse=true)
5×5 SparseMatrixCSC{Float32, Int64} with 12 stored entries:
 0.1   ⋅   0.1   ⋅    ⋅
 0.1  0.1   ⋅   0.1   ⋅
  ⋅   0.1  0.1   ⋅   0.1
  ⋅    ⋅   0.1  0.1   ⋅
  ⋅    ⋅    ⋅   0.1  0.1

References

Elsarraj, D.; Qisi, M. A.; Rodan, A.; Obeid, N.; Sharieh, A. and Faris, H. (2019). Demystifying echo state network with deterministic simple topologies. International Journal of Computational Science and Engineering 19, 407–417.