Models

Echo State Network

Constructor

ReservoirComputing.ESN — Type

ESN()

Train

Missing docstring.

Missing docstring for ESNtrain. Check Documenter's build log for details.

General function for training, used for all the AbstractReservoirComputers types with the exception of RMM which has the training method inside the constructor.

Predict

Missing docstring.

Missing docstring for ESNpredict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ESNpredict_h_steps. Check Documenter's build log for details.

For a full list of training and prediction methods for ESNs please refer to the User Guide ESN mods.

Double Activation Function Echo State Network

Constructor

Missing docstring.

Missing docstring for dafESN. Check Documenter's build log for details.

Predict

Missing docstring.

Missing docstring for dafESNpredict. Check Documenter's build log for details.

Missing docstring.

Missing docstring for dafESNpredict_h_steps. Check Documenter's build log for details.

Reservoir Computing with Cellular Automata

Constructors

Missing docstring.

Missing docstring for RECA_discrete. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RECA_TwoDim. Check Documenter's build log for details.

Predict

Missing docstring.

Missing docstring for RECAdirect_predict_discrete. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RECATDdirect_predict_discrete. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RECATD_predict_discrete. Check Documenter's build log for details.

Reservoir Memory Machine

Constructor

Missing docstring.

Missing docstring for RMM. Check Documenter's build log for details.

Predict

Missing docstring.

Missing docstring for RMMdirect_predict. Check Documenter's build log for details.

Gated Recurrent Unit ESN

Constructor

Missing docstring.

Missing docstring for GRUESN. Check Documenter's build log for details.

Predict

Missing docstring.

Missing docstring for GRUESNpredict. Check Documenter's build log for details.