Non Linear Algorithm

ReservoirComputing.NLADefault — Type

NLADefault()

Return the array untouched, default option.

ReservoirComputing.NLAT1 — Type

NLAT1()

Apply the $ \text{T}_1 $ transformation algorithm, as defined in [1] and [2].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).

[2] Pathak, Jaideep, et al. "Model-free prediction of large spatiotemporally chaotic systems from data: A reservoir computing approach." Physical review letters 120.2 (2018): 024102.

ReservoirComputing.NLAT2 — Type

NLAT2()

Apply the $ \text{T}_2 $ transformation algorithm, as defined in [1].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).

ReservoirComputing.NLAT3 — Type

NLAT3()

Apply the $ \text{T}_3 $ transformation algorithm, as defined in [1].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).