Special ESNs

These ESN are "special" in the fact that they have their own training methods. One of the future goals is to merge all the training methods into one. For all these models the constructor is the same, ESN.

Support Vector Echo State Machines

Train

ReservoirComputing.SVESMtrain — Function

SVESMtrain(svr::LIBSVM.AbstractSVR, esn::AbstractLeakyESN [, y_target::AbstractArray{Float64}])

Train the ESN using SVM regression, as described in [1].

[1] Shi, Zhiwei, and Min Han. "Support vector echo-state machine for chaotic time-series prediction." IEEE Transactions on Neural Networks 18.2 (2007): 359-372.

Predict

ReservoirComputing.SVESM_direct_predict — Function

SVESM_direct_predict(esn::AbstractLeakyESN, test_in::AbstractArray{Float64}, fitted_svr::LIBSVM.AbstractSVR)

Given the input data return the corresponding predicted output, as described in [1].

[1] Shi, Zhiwei, and Min Han. "Support vector echo-state machine for chaotic time-series prediction." IEEE Transactions on Neural Networks 18.2 (2007): 359-372.

ReservoirComputing.SVESMpredict — Function

SVESMpredict(esn::AbstractLeakyESN, predict_len::Int, fitted_svr::AbstractArray{Any})

Return the prediction for a given length of the constructed ESN struct using SVMs.

ReservoirComputing.SVESMpredict_h_steps — Function

SVESMpredict_h_steps(esn::AbstractLeakyESN, predict_len::Int, h_steps::Int, test_data::AbstractArray{Float64}, fitted_svr::LIBSVM.AbstractSVR)

Return the prediction for h steps ahead of the constructed ESN struct using SVMs.

Echo State Gaussian Processes

Train

ReservoirComputing.ESGPtrain — Function

ESGPtrain(esn::AbstractLeakyESN, mean::GaussianProcesses.Mean, kernel::GaussianProcesses.Kernel
[, lognoise::Float64, optimize::Bool, optimizer::Optim.AbstractOptimizer, y_target::AbstractArray{Float64})

Train the ESN using Gaussian Processes, as described in [1]

[1] Chatzis, Sotirios P., and Yiannis Demiris. "Echo state Gaussian process." IEEE Transactions on Neural Networks 22.9 (2011): 1435-1445.

Predict

ReservoirComputing.ESGPpredict — Function

ESGPpredict(esn::AbstractLeakyESN, predict_len::Int, gp::AbstractArray{Any})

Return the prediction for a given length of the constructed ESN struct using GPs.

ReservoirComputing.ESGPpredict_h_steps — Function

ESGPpredict_h_steps(esn::AbstractLeakyESN, predict_len::Int, h_steps::Int, test_data::AbstractArray{Float64}, gp::GaussianProcesses.GPE)

Return the prediction for h steps ahead of the constructed ESN struct using GPs.