Prediction Types

ReservoirComputing.Generative — Type

Generative(prediction_len)

This prediction methodology allows the models to produce an autonomous prediction, feeding the prediction into itself to generate the next step. The only parameter needed is the number of steps for the prediction.

ReservoirComputing.Predictive — Type

Predictive(prediction_data)

Given a set of labels as prediction_data this method of prediction will return the correspinding labels in a standard Machine Learning fashion.