States Modifications

Padding and Estension

ReservoirComputing.StandardStates — Type

StandardStates()

No modification of the states takes place, default option.

ReservoirComputing.ExtendedStates — Type

ExtendedStates()

The states are extended with the input data, for the training section, and the prediction data, during the prediction section. This is obtained with a vertical concatenation of the data and the states.

ReservoirComputing.PaddedStates — Type

PaddedStates(padding)
PaddedStates(;padding=1.0)

The states are padded with a chosen value. Usually this value is set to one. The padding is obtained through a vertical concatenation of the padding value and the states.

ReservoirComputing.PaddedExtendedStates — Type

PaddedExtendedStates(padding)
PaddedExtendedStates(;padding=1.0)

The states are extended with the training data or predicted data and subsequently padded with a chosen value. Usually the padding value is set to one. The padding and the extension are obtained through a vertical concatenation of the padding value, the data and the states.

Non Linear Transformations

ReservoirComputing.NLADefault — Type

NLADefault()

Returns the array untouched, default option.

ReservoirComputing.NLAT1 — Type

NLAT1()

Applies the $ \text{T}_1 $ transformation algorithm, as defined in [1] and [2].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).

[2] Pathak, Jaideep, et al. "Model-free prediction of large spatiotemporally chaotic systems from data: A reservoir computing approach." Physical review letters 120.2 (2018): 024102.

ReservoirComputing.NLAT2 — Type

NLAT2()

Apply the $ \text{T}_2 $ transformation algorithm, as defined in [1].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).

ReservoirComputing.NLAT3 — Type

NLAT3()

Apply the $ \text{T}_3 $ transformation algorithm, as defined in [1].

[1] Chattopadhyay, Ashesh, et al. "Data-driven prediction of a multi-scale Lorenz 96 chaotic system using a hierarchy of deep learning methods: Reservoir computing, ANN, and RNN-LSTM." (2019).