Training Algorithms

Linear Models

ReservoirComputing.StandardRidge — Type

StandardRidge([Type], [reg])

Returns a training method for train based on ridge regression. The equations for ridge regression are as follows:

\[\mathbf{w} = (\mathbf{X}^\top \mathbf{X} + \lambda \mathbf{I})^{-1} \mathbf{X}^\top \mathbf{y}\]

Arguments

Type: type of the regularization argument. Default is inferred internally, there's usually no need to tweak this

reg: regularization coefficient. Default is set to 0.0 (linear regression).

```

Gaussian Regression

Currently, v0.10, is unavailable.

Support Vector Regression

Support Vector Regression is possible using a direct call to LIBSVM regression methods. Instead of a wrapper, please refer to the use of LIBSVM.AbstractSVR in the original library.