Layers

Input Layers

ReservoirComputing.init_input_layer — Function

init_input_layer(res_size::Int, in_size::Int, sigma::Float64)

Return a weighted input layer matrix, with random non-zero elements drawn from $ [-\text{sigma}, \text{sigma}] $, as described in [1].

[1] Lu, Zhixin, et al. "Reservoir observers: Model-free inference of unmeasured variables in chaotic systems." Chaos: An Interdisciplinary Journal of Nonlinear Science 27.4 (2017): 041102.

ReservoirComputing.init_dense_input_layer — Function

init_dense_input_layer(res_size::Int, in_size::Int, sigma::Float64)

Return a fully connected input layer matrix, with random non-zero elements drawn from $ [-sigma, sigma] $.

ReservoirComputing.init_sparse_input_layer — Function

init_sparse_input_layer(res_size::Int, in_size::Int, sigma::Float64, sparsity::Float64)

Return a sparsely connected input layer matrix, with random non-zero elements drawn from $ [-sigma, sigma] $ and given sparsity.

ReservoirComputing.min_complex_input — Function

min_complex_input(res_size::Int, in_size::Int, weight::Float64)

Return a fully connected input layer matrix with the same weights and sign drawn from a Bernoulli distribution, as described in [1].

[1] Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.irrational_sign_input — Function

irrational_sign_input(res_size::Int, in_size::Int , weight::Float64 [, start::Int, irrational::Irrational])

Return a fully connected input layer matrix with the same weights and sign decided by the values of an irrational number, as described in [1] and [2].

[1] Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144. [2] Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs with regular jumps." Neural computation 24.7 (2012): 1822-1852.

Reservoirs

ReservoirComputing.init_reservoir_givendeg — Function

 init_reservoir_givendeg(res_size::Int, radius::Float64, degree::Int)

Return a reservoir matrix scaled by the radius value and with a given degree of connection.

ReservoirComputing.init_reservoir_givensp — Function

 init_reservoir_givensp(res_size::Int, radius::Float64, sparsity::Float64)

Return a reservoir matrix scaled by the radius value and with a given sparsity.

ReservoirComputing.pseudoSVD — Function

pseudoSVD(dim::Int,  max_value::Float64, sparsity::Float64 [, sorted::Bool, reverse_sort::Bool])

Return a reservoir matrix created using SVD as described in [1].

[1] Yang, Cuili, et al. "Design of polynomial echo state networks for time series prediction." Neurocomputing 290 (2018): 148-160.

ReservoirComputing.DLR — Function

DLR(res_size::Int, weight::Float64)

Return a Delay Line Reservoir matrix as described in [2].

[2] Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.DLRB — Function

DLRB(res_size::Int, weight::Float64, fb_weight::Float64)

Return a Delay Line Reservoir matrix with Backward connections as described in [2].

[2] Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.SCR — Function

SCR(res_size::Int, weight::Float64)

Return a Simple Cycle Reservoir Reservoir matrix as described in [2].

[2] Rodan, Ali, and Peter Tino. "Minimum complexity echo state network." IEEE transactions on neural networks 22.1 (2010): 131-144.

ReservoirComputing.CRJ — Function

CRJ(res_size::Int, cycle_weight::Float64, jump_weight::Float64, jump_size::Int)

Return a Cycle Reservoir with Jumps matrix as described in [2].

[2] Rodan, Ali, and Peter Tiňo. "Simple deterministically constructed cycle reservoirs with regular jumps." Neural computation 24.7 (2012): 1822-1852.

Reservoirs only usable for RECA models:

ReservoirComputing.ECA — Type

ECA(rule::Int, starting_val::AbstractArray{Int} [, generations::Int])

Creates an Elementary Cellular Automaton [1] struct given the rule and a starting vector.

[1] Wolfram, Stephen. A new kind of science. Vol. 5. Champaign, IL: Wolfram media, 2002.

ReservoirComputing.GameOfLife — Type

function GameOfLife(initial_state::AbstractArray{T}, generations::Int)

Given a starting matrix return the evolution for given generations of Conway's Game of Life, as described in [1].

[1] Gardner, Martin. “Mathematical games: The fantastic combinations of John Conway’s new solitaire game “life”.” Scientific American 223.4 (1970): 120-123.