Models

Echo State Network

Constructor

ReservoirComputing.ESN — Type

ESN(W::AbstractArray{T}, train_data::AbstractArray{T}, W_in::AbstractArray{T}
[, activation::Any, alpha::T, nla_type::NonLinearAlgorithm, extended_states::Bool])

Build an ESN struct given the input and reservoir matrices.

Train

ReservoirComputing.ESNtrain — Function

ESNtrain(esn::AbstractReservoirComputer, beta::Float64[, train_data])

Return the trained output layer using Ridge Regression.

ESNtrain(lm::LinearModel, esn::AbstractReservoirComputer[, train_data])

Return the trained output layer using an MLJLinearModels method built into a LinearSolver struct.

General function for training, used for all the AbstractReservoirComputers types with the exception of RMM which has the training method inside the constructor.

Predict

ReservoirComputing.ESNpredict — Function

ESNpredict(esn::AbstractLeakyESN, predict_len::Int, W_out::AbstractArray{Float64})

Return the prediction for a given length of the constructed ESN struct.

ReservoirComputing.ESNpredict_h_steps — Function

ESNpredict_h_steps(esn::AbstractLeakyESN, predict_len::Int, h_steps::Int,
test_data::AbstractArray{Float64}, W_out::AbstractArray{Float64})

Return the prediction for h steps ahead of the constructed ESN struct.

For a full list of training and prediction methods for ESNs please refer to the User Guide ESN mods.

Double Activation Function Echo State Network

Constructor

ReservoirComputing.dafESN — Type

dafESN(W::AbstractArray{T}, train_data::AbstractArray{T}, first_lambda::T,
second_lambda::T, W_in::AbstractArray{T} [, first_activation::Any,
second_activation::Any, alpha::T, nla_type::NonLinearAlgorithm, extended_states::Bool])

Build a double activation function ESN struct given the input and reservoir matrices, as described in [1].

[1]: Lun, Shu-Xian, et al. "A novel model of leaky integrator echo state network for time-series prediction." Neurocomputing 159 (2015): 58-66.

Predict

ReservoirComputing.dafESNpredict — Function

dafESNpredict(esn::AbstractLeakyESN, predict_len::Int, W_out::AbstractArray{Float64})

Return the prediction for a given length of the constructed dafESN struct.

ReservoirComputing.dafESNpredict_h_steps — Function

dafESNpredict_h_steps(esn::AbstractLeakyESN, predict_len::Int, h_steps::Int,
test_data::AbstractArray{Float64}, W_out::AbstractArray{Float64})

Return the prediction for h steps ahead of the constructed ESN struct.

Reservoir Computing with Cellular Automata

Constructors

ReservoirComputing.RECA_discrete — Type

RECA_discrete(train_data::AbstractArray{Int}, rule::Int, generations::Int, 
expansion_size::Int, permutations::Int [, nla_type])

Create a RECA struct as described in [1] and [2].

[1] Yilmaz, Ozgur. “Reservoir computing using cellular automata.” arXiv preprint arXiv:1410.0162 (2014). [2] Nichele, Stefano, and Andreas Molund. “Deep reservoir computing using cellular automata.” arXiv preprint arXiv:1703.02806 (2017).

ReservoirComputing.RECA_TwoDim — Type

RECA_TwoDim(train_data, res_size, generations, permutations [, nla_type])

Create the RECA_TwoDim struct for two-dimensional Cellular Automata Reservoir Computing, as described in [1].

[1] Yilmaz, Ozgur. “Reservoir computing using cellular automata.” arXiv preprint arXiv:1410.0162 (2014).

Predict

ReservoirComputing.RECAdirect_predict_discrete — Function

RECAdirect_predict_discrete(reca::AbstractReca, W_out::AbstractArray{Float64}, test_data::AbstractArray{Int})

Given the input data return the corresponding predicted output, as described in [1].

[1] Yilmaz, Ozgur. “Reservoir computing using cellular automata.” arXiv preprint arXiv:1410.0162 (2014).

ReservoirComputing.RECATDdirect_predict_discrete — Function

RECATDdirect_predict_discrete(reca::AbstractReca, W_out::AbstractArray{Float64}, test_data::AbstractArray{Int})

Given the input data return the corresponding predicted output, as described in [1].

[1] Yilmaz, Ozgur. “Reservoir computing using cellular automata.” arXiv preprint arXiv:1410.0162 (2014).

ReservoirComputing.RECATD_predict_discrete — Function

RECATD_predict_discrete(reca, predict_len::Int, W_out::AbstractArray{Float64})

Return the prediction for a given length of the constructed RECA_TwoDim struct.

Reservoir Memory Machine

Constructor

ReservoirComputing.RMM — Type

RMM(W::AbstractArray{T}, in_data::AbstractArray{T}, out_data::AbstractArray{T}, W_in::AbstractArray{T},
memory_size::Int, beta::Float64 [, activation::Any, alpha::T, nla_type::NonLinearAlgorithm, extended_states::Bool])

Create a Reservoir Memory Machine struct, as described in [1].

[1] Paaßen, Benjamin, and Alexander Schulz. "Reservoir memory machines." arXiv preprint arXiv:2003.04793 (2020).

Predict

ReservoirComputing.RMMdirect_predict — Function

RMMdirect_predict(rmm::AbstractReservoirMemoryMachine, input)

Given the input data return the corresponding predicted output, as described in [1].

[1] Paaßen, Benjamin, and Alexander Schulz. "Reservoir memory machines." arXiv preprint arXiv:2003.04793 (2020).

Gated Recurrent Unit ESN

Constructor

ReservoirComputing.GRUESN — Type

GRUESN(W::AbstractArray{T}, train_data::AbstractArray{T}, W_in::AbstractArray{T} [, gates_weight::T, activation::Any, alpha::T, nla_type::NonLinearAlgorithm, extended_states::Bool])

Return a Gated Recurrent Unit [1] ESN struct.

[1] Cho, Kyunghyun, et al. “Learning phrase representations using RNN encoder-decoder for statistical machine translation.” arXiv preprint arXiv:1406.1078 (2014).

Predict

ReservoirComputing.GRUESNpredict — Function

GRUESNpredict(esn::AbstractGRUESN, predict_len::Int, W_out::AbstractArray{Float64})

Return the prediction for a given length of the constructed GRUESN.